pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md2ti84"
version = "0.1.0"
description = "Convert Markdown notes to TI-84 CE calculator images"
requires-python = ">=3.10"
dependencies = [
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "jinja2>=3.1",
    "pdf2image>=1.16",
    "Pillow>=10.0",
]

[project.scripts]
md2ti84 = "md2ti84.cli:main"

[tool.md2ti84]
# Target TI-84 CE image var dimensions
image_width = 265
image_height = 165
# LaTeX font size in pt
font_size = 7
# Page margin in pt
margin = 4
# PDF render DPI before downscaling
render_dpi = 300
# LaTeX engine: "lualatex" or "pdflatex"
latex_engine = "lualatex"
# Estimated max lines per page (tune this empirically)
max_lines_per_page = 18
# Grayscale by default (crisper text, smaller files)
grayscale = true
